requirements.txt
torch==2.1.0
torchvision==0.16.0
opencv-python==4.8.0.74
scikit-learn==1.3.0
Pillow==10.0.0
tqdm==4.66.1
matplotlib==3.8.0
numpy==1.24.3
